ci: Test using seatd instead of direct launcher
While the commonly used Weston launchers are weston-launch and launcher-logind,
the direct backend was used in CI out of convenience, and due to logind being a
bit cumbersome to get to work in a CI environment.

The new libseat launcher can be used with seatd as well as logind. seatd is easy
to start in a CI environment, allowing us to test the libseat launcher codepath
instead of the less user relevant direct launcher.

This also prepares us for the future intended removal of non-libseat launchers.

Signed-off-by: Kenny Levinsen <kl@kl.wtf>